♻️ Refactor Trips structure to match Notes
